+++ src/mcfacts/objects/actors/__init__.py
@@ -1,0 +1,3 @@
+from .reclassify_objects import *
+from .binary_bh import *
+from .single_bh import *
