+++ src/mcfacts/objects/actors/binary_bh.py
@@ -1,4 +1,3 @@
-from abc import ABC, abstractmethod
 from copy import deepcopy
 
 import numpy as np
@@ -6,390 +5,15 @@ from numpy.random import Generator
 
 from mcfacts.inputs.settings_manager import SettingsManager
 from mcfacts.mcfacts_random_state import uuid_provider
-from mcfacts.objects.agn_disk import AGNDisk
 from mcfacts.objects.agn_object_array import AGNBlackHoleArray, AGNBinaryBlackHoleArray, AGNMergedBlackHoleArray
-from mcfacts.objects.filing_cabinet import FilingCabinet
-from mcfacts.physics import migration, accretion, eccentricity, disk_capture, dynamics, gw, feedback
+from mcfacts.objects.galaxy import FilingCabinet, AGNDisk
+from mcfacts.objects.timeline import TimelineActor
+from mcfacts.physics import migration, eccentricity, dynamics, gw
 from mcfacts.physics.binary import evolve, merge, formation
 from mcfacts.physics.gw import gw_strain_freq
-from mcfacts.setup import setupdiskblackholes
 
 
-class SimulationActor(ABC):
-    """
-    A base class representing a generic simulation actor in the simulation framework.
-
-    Simulation actors define behaviors or actions to be performed during the simulation,
-    and each actor is identified by a name and utilizes specific settings.
-
-    Attributes:
-        name (str): The name of the simulation actor.
-        settings (SettingsManager): A manager for accessing configuration settings. Defaults to the base `SettingsManager`.
-
-    Methods:
-        perform(timestep: int, filing_cabinet: FilingCabinet, randomState: RandomState):
-            Abstract method to be implemented by subclasses. Defines the behavior of the simulation actor for a given timestep.
-
-        __str__():
-            Returns a string representation of the simulation actor, including its name and type.
-    """
-
-    def __init__(self, name: str, settings: SettingsManager = SettingsManager()):
-        """
-        Initializes a SimulationActor instance.
-
-        Args:
-            name (str): The name of the simulation actor.
-            settings (SettingsManager, optional): The settings manager to retrieve configurations for the actor.
-                Defaults to an empty `SettingsManager`.
-        """
-        self.name: str = name
-        self.settings: SettingsManager = settings
-
-    @abstractmethod
-    def perform(self, timestep: int, timestep_length: float, time_passed: float, filing_cabinet: FilingCabinet, agn_disk: AGNDisk, random_generator: Generator):
-        """
-        Abstract method to define the behavior of the simulation actor during a timestep.
-
-        Subclasses must implement this method to specify actions to be taken in the simulation.
-
-        Args:
-            timestep (int): The current simulation timestep.
-            timestep_length (float): The duration of a timestep in years.
-            time_passed (float): Time passed in years.
-            filing_cabinet (FilingCabinet): A container for accessing and modifying simulation data arrays.
-            agn_disk (AGNDisk): An object containing the generated properties of the AGN disk.
-            random_generator (Generator): A random number generator for stochastic behaviors.
-
-        Returns:
-            NotImplemented: Must be overridden by subclasses.
-        """
-        return NotImplemented
-
-    def __str__(self):
-        """
-        Returns a string representation of the simulation actor.
-
-        The string includes the actor's name and its class type.
-
-        Returns:
-            str: A formatted string representation of the simulation actor.
-        """
-        return f"{self.name} ({type(self)})"
-
-
-class ReclassifyDiskObjects(SimulationActor):
-    """
-    This simulation actor looks at the initial seeded array of objects and distributes them
-    based on different classifications.
-
-    Attributes:
-        name (str): Optional. The name of the simulation actor. Defaults to "Reclassify Disk Objects".
-        settings (SettingsManager): A manager for accessing configuration settings. Defaults to an empty `SettingsManager`.
-
-    Methods:
-        perform(timestep, filing_cabinet, randomState):
-            Classifies black holes based on their semi-major axis and orbital angular momentum
-            and stores the results in the filing cabinet.
-
-    Notes:
-        - Black holes with a semi-major axis smaller than `settings.inner_disk_outer_radius` are classified as inner disk black holes.
-        - Prograde black holes are identified by a positive orbital angular momentum (`orb_ang_mom > 0`).
-        - Retrograde black holes are identified by a negative orbital angular momentum (`orb_ang_mom < 0`).
-        - At the end of this process, there should be no unclassified black holes left in the original array (`settings.bh_array_name`).
-    """
-
-    def __init__(self, name: str = None, settings: SettingsManager = SettingsManager()):
-        """
-        Initializes the `ReclassifyDiskObjects` simulation actor.
-
-        Args:
-            name (str): Optional. The name of the actor. Defaults to "Reclassify Disk Objects".
-            settings (SettingsManager): A settings manager instance. Defaults to base instance of `SettingsManager`.
-        """
-        super().__init__("Reclassify Disk Objects" if name is None else name, settings)
-
-    def perform(self, timestep: int, timestep_length: float, time_passed: float, filing_cabinet: FilingCabinet, agn_disk: AGNDisk, random_generator: Generator):
-        """
-        Performs the classification of objects into separate categories and updates the filing cabinet.
-
-        Args:
-            timestep (int): The current simulation timestep.
-            timestep_length (float): The duration of a timestep in years.
-            time_passed (float): Time passed in years.
-            filing_cabinet (FilingCabinet): A container for managing and storing categorized black holes.
-            agn_disk (AGNDisk): An object containing the generated properties of the AGN disk.
-            random_generator (Generator): A random number generator for stochastic behaviors.
-
-        Black Hole Classification Steps:
-            1. Inner Disk Black Holes:
-               - Black holes with a semi-major axis (`orb_a`) smaller than `settings.inner_disk_outer_radius`.
-               - Moved to a separate array (`settings.bh_inner_disk_array_name`).
-
-            2. Prograde Black Holes:
-               - Black holes with a positive orbital angular momentum (`orb_ang_mom > 0`).
-               - Moved to a separate array (`settings.bh_prograde_array_name`).
-
-            3. Retrograde Black Holes:
-               - Black holes with a negative orbital angular momentum (`orb_ang_mom < 0`).
-               - Moved to a separate array (`settings.bh_retrograde_array_name`).
-
-        Notes:
-            - At the end of this method, the original black hole array (`settings.bh_array_name`) should be empty.
-        """
-        sm = self.settings
-
-        # Black Holes, get array from filing cabinet or create an empty one if it doesn't
-
-        blackholes = filing_cabinet.get_array(sm.bh_array_name, AGNBlackHoleArray, True)
-
-        # Inner disk black holes
-
-        inner_disk_blackholes = blackholes.copy()
-
-        # Create a mask for all black holes with a semi-major axis smaller than the defined outer radius
-        inner_disk_bh_ids = blackholes.unique_id[blackholes.orb_a < sm.inner_disk_outer_radius]
-
-        # Apply the mask the two lists
-        inner_disk_blackholes.keep_only(inner_disk_bh_ids)
-        blackholes.remove_all(inner_disk_bh_ids)
-
-        # Add inner disk black holes to separate filing cabinet array.
-        filing_cabinet.create_or_append_array(sm.bh_inner_disk_array_name, inner_disk_blackholes)
-
-        # Prograde black holes
-
-        prograde_blackholes = blackholes.copy()
-
-        # Find prograde BH orbiters. Identify BH with orb. ang mom > 0 (orb_ang_mom is only ever +1 or -1)
-        prograde_bh_ids = blackholes.unique_id[blackholes.orb_ang_mom > 0]
-
-        # Apply the mask
-        prograde_blackholes.keep_only(prograde_bh_ids)
-        blackholes.remove_all(prograde_bh_ids)
-
-        # Add prograde disk black holes to separate filing cabinet array.
-        filing_cabinet.create_or_append_array(sm.bh_prograde_array_name, prograde_blackholes)
-
-        # Retrograde black holes
-
-        retrograde_blackholes = blackholes.copy()
-
-        # Find retrograde black holes
-        retrograde_bh_ids = blackholes.unique_id[blackholes.orb_ang_mom < 0]
-
-        # Apply the mask
-        retrograde_blackholes.keep_only(retrograde_bh_ids)
-        blackholes.remove_all(retrograde_bh_ids)
-
-        # Add prograde disk black holes to separate filing cabinet array.
-        filing_cabinet.create_or_append_array(sm.bh_retrograde_array_name, retrograde_blackholes)
-
-        # At the end of this, we should have no black holes in the sm.bh_array_name array,
-        # it can be used in the future to seed new black holes before classification is run.
-
-        # TODO: Stars
-
-
-class ProgradeBlackHoleDynamics(SimulationActor):
-    def __init__(self, name: str = None, settings: SettingsManager = SettingsManager()):
-        super().__init__("Prograde Black Hole Dynamics" if name is None else name, settings)
-
-    def perform(self, timestep: int, timestep_length: float, time_passed: float, filing_cabinet: FilingCabinet, agn_disk: AGNDisk, random_generator: Generator):
-        sm = self.settings
-
-        # Check to make sure the array exists in the filing cabinet
-        if sm.bh_prograde_array_name not in filing_cabinet:
-            return
-
-        blackholes_pro = filing_cabinet.get_array(sm.bh_prograde_array_name, AGNBlackHoleArray)
-
-        # Migrate
-        # First if feedback present, find ratio of feedback heating torque to migration torque
-        if sm.flag_thermal_feedback > 0:
-            ratio_heat_mig_torques = feedback.feedback_bh_hankla(
-                blackholes_pro.orb_a,
-                agn_disk.disk_surface_density,
-                agn_disk.disk_opacity,
-                sm.disk_bh_eddington_ratio,
-                sm.disk_alpha_viscosity,
-                sm.disk_radius_outer)
-        else:
-            ratio_heat_mig_torques = np.ones(len(blackholes_pro.orb_a))
-
-        # Migrate as usual
-        blackholes_pro.orb_a = migration.type1_migration_single(
-            sm.smbh_mass,
-            blackholes_pro.orb_a,
-            blackholes_pro.mass,
-            blackholes_pro.orb_ecc,
-            sm.disk_bh_pro_orb_ecc_crit,
-            agn_disk.disk_surface_density,
-            agn_disk.disk_aspect_ratio,
-            ratio_heat_mig_torques,
-            sm.disk_radius_trap,
-            sm.disk_radius_outer,
-            sm.timestep_duration_yr
-        )
-
-        # Check for orb_a unphysical
-        bh_pro_id_num_unphysical_a = blackholes_pro.unique_id[blackholes_pro.orb_a == 0.0]
-        blackholes_pro.remove_all(bh_pro_id_num_unphysical_a)
-
-        # Accrete
-        blackholes_pro.mass = accretion.change_bh_mass(
-            blackholes_pro.mass,
-            sm.disk_bh_eddington_ratio,
-            sm.disk_bh_eddington_mass_growth_rate,
-            sm.timestep_duration_yr
-        )
-
-        # Spin up
-        blackholes_pro.spin = accretion.change_bh_spin_magnitudes(
-            blackholes_pro.spin,
-            sm.disk_bh_eddington_ratio,
-            sm.disk_bh_torque_condition,
-            sm.timestep_duration_yr,
-            blackholes_pro.orb_ecc,
-            sm.disk_bh_pro_orb_ecc_crit,
-        )
-
-        # Torque spin angle
-        blackholes_pro.spin_angle = accretion.change_bh_spin_angles(
-            blackholes_pro.spin_angle,
-            sm.disk_bh_eddington_ratio,
-            sm.disk_bh_torque_condition,
-            sm.disk_bh_spin_resolution_min,
-            sm.timestep_duration_yr,
-            blackholes_pro.orb_ecc,
-            sm.disk_bh_pro_orb_ecc_crit
-        )
-
-        # Damp BH orbital eccentricity
-        blackholes_pro.orb_ecc = eccentricity.orbital_ecc_damping(
-            sm.smbh_mass,
-            blackholes_pro.orb_a,
-            blackholes_pro.mass,
-            agn_disk.disk_surface_density,
-            agn_disk.disk_aspect_ratio,
-            blackholes_pro.orb_ecc,
-            sm.timestep_duration_yr,
-            sm.disk_bh_pro_orb_ecc_crit,
-        )
-
-        if sm.flag_dynamic_enc > 0:
-            blackholes_pro.orb_a, blackholes_pro.orb_ecc = dynamics.circular_singles_encounters_prograde(
-                sm.smbh_mass,
-                blackholes_pro.orb_a,
-                blackholes_pro.mass,
-                blackholes_pro.orb_ecc,
-                sm.timestep_duration_yr,
-                sm.disk_bh_pro_orb_ecc_crit,
-                sm.delta_energy_strong,
-            )
-
-        # After this time period, was there a disk capture via orbital grind-down?
-        # To do: What eccentricity do we want the captured BH to have? Right now ecc=0.0? Should it be ecc<h at a?
-        # Assume 1st gen BH captured and orb ecc =0.0
-        # To do: Bias disk capture to more massive BH!
-        capture = time_passed % sm.capture_time_yr
-        if capture == 0:
-            disk_bh_num = 1
-
-            bh_orb_a_captured = setupdiskblackholes.setup_disk_blackholes_location(disk_bh_num, sm.disk_radius_capture_outer, sm.disk_inner_stable_circ_orb)
-            bh_mass_captured = setupdiskblackholes.setup_disk_blackholes_masses(disk_bh_num, sm.nsc_imf_bh_mode, sm.nsc_imf_bh_mass_max, sm.nsc_imf_bh_powerlaw_index, sm.mass_pile_up)
-            bh_spin_captured = setupdiskblackholes.setup_disk_blackholes_spins(disk_bh_num, sm.nsc_bh_spin_dist_mu, sm.nsc_bh_spin_dist_sigma)
-            bh_spin_angle_captured = setupdiskblackholes.setup_disk_blackholes_spin_angles(disk_bh_num, bh_spin_captured)
-
-            unique_ids = np.array([uuid_provider(random_generator) for _ in range(disk_bh_num)])
-
-            captured_prograde_bh = AGNBlackHoleArray(
-                unique_id=unique_ids,
-                mass=bh_mass_captured,
-                spin=bh_spin_captured,
-                spin_angle=bh_spin_angle_captured,
-                orb_a=bh_orb_a_captured,
-                orb_inc=np.full(disk_bh_num, 0.0),
-                orb_ecc=np.full(disk_bh_num, 0.0),
-                orb_ang_mom=np.ones(disk_bh_num),
-                orb_arg_periapse=np.full(disk_bh_num, -1.5),
-                gen=np.ones(disk_bh_num)
-            )
-
-            blackholes_pro.add_objects(captured_prograde_bh)
-
-        # Test if any BH or BBH are in the danger-zone (<mininum_safe_distance, default =50r_g) from SMBH.
-        # Potential EMRI/BBH EMRIs.
-        # Find prograde BH in inner disk. Define inner disk as <=50r_g.
-        # Since a 10Msun BH will decay into a 10^8Msun SMBH at 50R_g in ~38Myr and decay time propto a^4.
-        # e.g at 25R_g, decay time is only 2.3Myr.
-
-        # Remove the pros if it is in the inner disk and add it to the inner disk array.
-        pro_bh_inner_disk_ids = blackholes_pro.unique_id[blackholes_pro.orb_a < sm.inner_disk_outer_radius]
-
-        inner_pros = blackholes_pro.copy()
-        inner_pros.keep_only(pro_bh_inner_disk_ids)
-        filing_cabinet.create_or_append_array(sm.bh_inner_disk_array_name, inner_pros)
-
-        blackholes_pro.remove_all(pro_bh_inner_disk_ids)
-
-
-class RetrogradeBlackholeDynamics(SimulationActor):
-    def __init__(self, name: str = None, settings: SettingsManager = SettingsManager()):
-        super().__init__("Retrograde Black Hole Dynamics" if name is None else name, settings)
-
-    def perform(self, timestep: int, timestep_length: float, time_passed: float, filing_cabinet: FilingCabinet, agn_disk: AGNDisk, random_generator: Generator):
-        sm = self.settings
-
-        # Check to make sure the array exists in the filing cabinet
-        if sm.bh_retrograde_array_name not in filing_cabinet:
-            return
-
-        blackholes_retro = filing_cabinet.get_array(sm.bh_retrograde_array_name, AGNBlackHoleArray)
-
-        orb_ecc, orb_a, orb_inc = disk_capture.retro_bh_orb_disk_evolve(
-            sm.smbh_mass,
-            blackholes_retro.mass,
-            blackholes_retro.orb_a,
-            blackholes_retro.orb_ecc,
-            blackholes_retro.orb_inc,
-            blackholes_retro.orb_arg_periapse,
-            sm.disk_inner_stable_circ_orb,
-            agn_disk.disk_surface_density,
-            timestep_length
-        )
-
-        # Update retro black hole object array attributes
-        blackholes_retro.orb_ecc = np.array(orb_ecc)
-        blackholes_retro.orb_a = np.array(orb_a)
-        blackholes_retro.orb_inc = np.array(orb_inc)
-
-        # Check for unphysical orbital eccentricities
-        bh_retro_id_num_unphysical_ecc = blackholes_retro.unique_id[blackholes_retro.orb_ecc >= 1.]
-        blackholes_retro.remove_all(bh_retro_id_num_unphysical_ecc)
-
-        # Remove the retro if it is in the inner disk and add it to the inner disk array.
-        retro_bh_inner_disk_ids = blackholes_retro.unique_id[blackholes_retro.orb_a < sm.inner_disk_outer_radius]
-
-        inner_retros = blackholes_retro.copy()
-        inner_retros.keep_only(retro_bh_inner_disk_ids)
-        filing_cabinet.create_or_append_array(sm.bh_inner_disk_array_name, inner_retros)
-
-        blackholes_retro.remove_all(retro_bh_inner_disk_ids)
-
-        # Remove any retros that have flipped prograde and add them to the prograde black holes array
-        inc_threshhold = 5.0 * np.pi / 180.0
-
-        bh_id_num_flip_to_pro = blackholes_retro.unique_id[np.where((np.abs(blackholes_retro.orb_inc) <= inc_threshhold) | (blackholes_retro.orb_ecc == 0.0))]
-
-        blackholes_pro = blackholes_retro.copy()
-        blackholes_pro.keep_only(bh_id_num_flip_to_pro)
-        filing_cabinet.create_or_append_array(sm.bh_prograde_array_name, blackholes_pro)
-
-        blackholes_retro.remove_all(bh_id_num_flip_to_pro)
-
-
-class BinaryBlackHoleDynamics(SimulationActor):
+class BinaryBlackHoleDynamics(TimelineActor):
     def __init__(self, name: str = None, settings: SettingsManager = SettingsManager()):
         super().__init__("Binary Black Hole Dynamics" if name is None else name, settings)
 
@@ -729,7 +353,7 @@ class BinaryBlackHoleDynamics(SimulationActor):
                 filing_cabinet.create_or_append_array(sm.bh_prograde_array_name, new_blackholes)
 
 
-class BinaryBlackHoleFormation(SimulationActor):
+class BinaryBlackHoleFormation(TimelineActor):
     def __init__(self, name: str = None, settings: SettingsManager = SettingsManager()):
         super().__init__("Binary Black Hole Formation" if name is None else name, settings)
 
@@ -812,7 +436,7 @@ class BinaryBlackHoleFormation(SimulationActor):
             print(f"Binaries created: {primary_ids.size}")
 
 
-class BreakupBinaryBlackHoles(SimulationActor):
+class BreakupBinaryBlackHoles(TimelineActor):
     def __init__(self, name: str = None, settings: SettingsManager = SettingsManager()):
         super().__init__("Breakup Binary Black Holes" if name is None else name, settings)
