+++ src/mcfacts/objects/actors/reclassify_objects.py
@@ -1,0 +1,121 @@
+from numpy.random import Generator
+
+from mcfacts.inputs.settings_manager import SettingsManager
+from mcfacts.objects.agn_object_array import AGNBlackHoleArray
+from mcfacts.objects.galaxy import AGNDisk, FilingCabinet
+from mcfacts.objects.timeline import TimelineActor
+
+
+class ReclassifyDiskObjects(TimelineActor):
+    """
+    This simulation actor looks at the initial seeded array of objects and distributes them
+    based on different classifications.
+
+    Attributes:
+        name (str): Optional. The name of the simulation actor. Defaults to "Reclassify Disk Objects".
+        settings (SettingsManager): A manager for accessing configuration settings. Defaults to an empty `SettingsManager`.
+
+    Methods:
+        perform(timestep, filing_cabinet, randomState):
+            Classifies black holes based on their semi-major axis and orbital angular momentum
+            and stores the results in the filing cabinet.
+
+    Notes:
+        - Black holes with a semi-major axis smaller than `settings.inner_disk_outer_radius` are classified as inner disk black holes.
+        - Prograde black holes are identified by a positive orbital angular momentum (`orb_ang_mom > 0`).
+        - Retrograde black holes are identified by a negative orbital angular momentum (`orb_ang_mom < 0`).
+        - At the end of this process, there should be no unclassified black holes left in the original array (`settings.bh_array_name`).
+    """
+
+    def __init__(self, name: str = None, settings: SettingsManager = SettingsManager()):
+        """
+        Initializes the `ReclassifyDiskObjects` simulation actor.
+
+        Args:
+            name (str): Optional. The name of the actor. Defaults to "Reclassify Disk Objects".
+            settings (SettingsManager): A settings manager instance. Defaults to base instance of `SettingsManager`.
+        """
+        super().__init__("Reclassify Disk Objects" if name is None else name, settings)
+
+    def perform(self, timestep: int, timestep_length: float, time_passed: float, filing_cabinet: FilingCabinet, agn_disk: AGNDisk, random_generator: Generator):
+        """
+        Performs the classification of objects into separate categories and updates the filing cabinet.
+
+        Args:
+            timestep (int): The current simulation timestep.
+            timestep_length (float): The duration of a timestep in years.
+            time_passed (float): Time passed in years.
+            filing_cabinet (FilingCabinet): A container for managing and storing categorized black holes.
+            agn_disk (AGNDisk): An object containing the generated properties of the AGN disk.
+            random_generator (Generator): A random number generator for stochastic behaviors.
+
+        Black Hole Classification Steps:
+            1. Inner Disk Black Holes:
+               - Black holes with a semi-major axis (`orb_a`) smaller than `settings.inner_disk_outer_radius`.
+               - Moved to a separate array (`settings.bh_inner_disk_array_name`).
+
+            2. Prograde Black Holes:
+               - Black holes with a positive orbital angular momentum (`orb_ang_mom > 0`).
+               - Moved to a separate array (`settings.bh_prograde_array_name`).
+
+            3. Retrograde Black Holes:
+               - Black holes with a negative orbital angular momentum (`orb_ang_mom < 0`).
+               - Moved to a separate array (`settings.bh_retrograde_array_name`).
+
+        Notes:
+            - At the end of this method, the original black hole array (`settings.bh_array_name`) should be empty.
+        """
+        sm = self.settings
+
+        # Black Holes, get array from filing cabinet or create an empty one if it doesn't
+
+        blackholes = filing_cabinet.get_array(sm.bh_array_name, AGNBlackHoleArray, True)
+
+        # Inner disk black holes
+
+        inner_disk_blackholes = blackholes.copy()
+
+        # Create a mask for all black holes with a semi-major axis smaller than the defined outer radius
+        inner_disk_bh_ids = blackholes.unique_id[blackholes.orb_a < sm.inner_disk_outer_radius]
+
+        # Apply the mask the two lists
+        inner_disk_blackholes.keep_only(inner_disk_bh_ids)
+        blackholes.remove_all(inner_disk_bh_ids)
+
+        # Add inner disk black holes to separate filing cabinet array.
+        filing_cabinet.create_or_append_array(sm.bh_inner_disk_array_name, inner_disk_blackholes)
+
+        # Prograde black holes
+
+        prograde_blackholes = blackholes.copy()
+
+        # Find prograde BH orbiters. Identify BH with orb. ang mom > 0 (orb_ang_mom is only ever +1 or -1)
+        prograde_bh_ids = blackholes.unique_id[blackholes.orb_ang_mom > 0]
+
+        # Apply the mask
+        prograde_blackholes.keep_only(prograde_bh_ids)
+        blackholes.remove_all(prograde_bh_ids)
+
+        # Add prograde disk black holes to separate filing cabinet array.
+        filing_cabinet.create_or_append_array(sm.bh_prograde_array_name, prograde_blackholes)
+
+        # Retrograde black holes
+
+        retrograde_blackholes = blackholes.copy()
+
+        # Find retrograde black holes
+        retrograde_bh_ids = blackholes.unique_id[blackholes.orb_ang_mom < 0]
+
+        # Apply the mask
+        retrograde_blackholes.keep_only(retrograde_bh_ids)
+        blackholes.remove_all(retrograde_bh_ids)
+
+        # Add prograde disk black holes to separate filing cabinet array.
+        filing_cabinet.create_or_append_array(sm.bh_retrograde_array_name, retrograde_blackholes)
+
+        self.log("Objects Reclassified")
+
+        # At the end of this, we should have no black holes in the sm.bh_array_name array,
+        # it can be used in the future to seed new black holes before classification is run.
+
+        # TODO: Stars
