+++ src/mcfacts/objects/populators/__init__.py
@@ -1,0 +1,1 @@
+from .base_populators import *
